# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tech Job API Documentation",
    "version": "1.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [],
  "paths": {
    "/api/agent": {
      "post": {
        "summary": "แชทกับ AI (Database-Aware Agent)",
        "description": "ส่งคำถามเพื่อคุยกับระบบ AI อัจฉริยะที่รู้จักข้อมูลใบงานและคลังวัสดุภายในระบบ",
        "tags": [
          "Agent"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "question"
                ],
                "properties": {
                  "question": {
                    "type": "string",
                    "description": "คำถามที่ต้องการถาม AI",
                    "example": "มีงานอะไรที่ยังค้างอยู่บ้าง?"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "คำตอบจาก AI",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "answer": {
                      "type": "string",
                      "description": "คำตอบที่ประมวลผลแล้ว"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/export": {
      "get": {
        "summary": "Export jobs to CSV",
        "tags": [
          "Jobs"
        ],
        "description": "Exports a list of jobs to a CSV formatted file. Requires Admin or Manager role.",
        "responses": {
          "200": {
            "description": "CSV file contents"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "summary": "Get all jobs",
        "tags": [
          "Jobs"
        ],
        "description": "Returns a list of all jobs with related data.",
        "responses": {
          "200": {
            "description": "Array of jobs"
          }
        }
      },
      "post": {
        "summary": "Create a new job",
        "tags": [
          "Jobs"
        ],
        "description": "Creates a new job. Requires Admin or Manager role.",
        "responses": {
          "200": {
            "description": "The newly created job"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/reports": {
      "get": {
        "summary": "Get all reports",
        "tags": [
          "Reports"
        ],
        "description": "Returns a list of reports. Configurable via status and type queries.",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of reports"
          }
        }
      },
      "post": {
        "summary": "Create a new report",
        "tags": [
          "Reports"
        ],
        "description": "Create a report or issue. Requires title, type, and reporterId.",
        "responses": {
          "201": {
            "description": "The created report"
          },
          "400": {
            "description": "Missing required fields"
          }
        }
      }
    },
...
  },
...
}